+++ source/benchmarks/memory_benchmark/implementations/l0/stream_memory_l0.cpp
@@ -45,7 +45,7 @@ static TestResult run(const StreamMemoryArguments &arguments, Statistics &statis
     const size_t elementSize = useDoubles ? sizeof(double) : sizeof(float);
     const size_t fillValue = 313u;
     const int32_t scalarValue = -999;
-    const uint32_t gws = arguments.size / elementSize;
+    const uint32_t gws = static_cast<uint32_t>(arguments.size / elementSize);
     const uint64_t timerResolution = levelzero.getTimerResoultion(levelzero.device);
 
     // Create module
